Cleanup SearchField
